undo file separation
