+++ src/annbatch/samplers/__init__.py
@@ -1,5 +1,4 @@
-from ._chunk_sampler import ChunkSampler
-from ._distributed_sampler import DistributedSampler
+from ._chunk_sampler import ChunkSampler, DistributedSampler
 from ._random_sampler import RandomSampler
 from ._sequential_sampler import SequentialSampler
